+++ include/tepp/rt_export.h
@@ -54,7 +54,7 @@ namespace te
 				return *this;
 			}
 
-			const_iterator& operator+=(int inc) noexcept {
+			const_iterator& operator+=(int64_t inc) noexcept {
 				this->index += inc;
 				return *this;
 			}
